+++ packages/web-components/src/components/ic-chip/ic-chip.types.ts
@@ -1,3 +1,1 @@
-export type IcChipSizes = "default" | "large" | "small";
-
 export type IcChipAppearance = "filled" | "outline";