+++ packages/web-components/src/components/ic-empty-state/ic-empty-state.types.ts
@@ -1,0 +1,1 @@
+export type IcEmptyStateAlignment = "left" | "center" | "right";